pyproject.toml
[project]
name = "ai-chat"
version = "0.1.0"
description = ""
authors = [
    {name = "Acejkee",email = "belousow.ilja@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "openai (>=1.69.0,<2.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "channels (>=4.2.1,<5.0.0)",
    "django-autoslug (>=1.9.9,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-timezone-field (<8.0) ; python_version >= \"3.10\" and python_version < \"4.0\"",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
